+++ apps/admin/src/pages/Monitoring.tsx
@@ -3,14 +3,14 @@ import { useState } from "react";
 import RumTab from "../features/monitoring/RumTab";
 import RateLimitsTab from "../features/monitoring/RateLimitsTab";
 import CacheTab from "../features/monitoring/CacheTab";
-import AuditLog from "./AuditLog";
+import AuditLogTab from "../features/monitoring/AuditLogTab";
 import Jobs from "./Jobs";
 
 const tabs = [
   { id: "rum", label: "RUM", component: <RumTab /> },
   { id: "rate-limits", label: "Rate limits", component: <RateLimitsTab /> },
   { id: "cache", label: "Cache", component: <CacheTab /> },
-  { id: "audit-log", label: "Audit log", component: <AuditLog /> },
+  { id: "audit-log", label: "Audit log", component: <AuditLogTab /> },
   { id: "jobs", label: "Jobs", component: <Jobs /> },
 ] as const;
